Improve error message for invalid JSON configuration
If the FE/NC servers were passed an invalid JSON config, they reported "Configuration file config.json not found".
This was due to the readObject function returning NULL on any error, and readConfiguration interpreting it as a file not found.

Add a check to readObject to see if the file exists and add logging on error. This catches issues with invalid JSON configs and file read errors but doesn't log for missing files.
Change the the error reporting in readConfiguration to "cannot load configuration file FILE (missing/invalid?)".