Get data acquire progress via the httpRequest data event
Instead of using pipe, use a data event with a callback for httpRequest, to record the current progress.

This gives a more efficient transfer size update as it avoids needing the scan the data folder.

Add a data source type of 'rsync' or 'http' to the data sources which is used to check if dataSizeProgress should be called.