Add support for pruning of external/unmanaged data
This is primarily useful for our integration with Globus. As we store data transferred from Globus in datalink structure as a separate data source,
datalink needs to know how to manage it. This adds support for scanning through data and any unmanaged data sources without entries can be pruned as required.

This adds the following configuration options:

data_max_days: maximum number of days to keep files
data_prune_mins: how often to run the prune data task in minutes