DataLink - Fix files unpacking before they are completed
Use file.end() instead of file.close() for createWriteStream in httpRequest - both work but file.end seems to be more common.

Don't resolve until the createWriteStream emits the close event. This fixes an issue with the file not being completed when we resolve right after file.close/file.end.
This only seems to affect node version < 18.