DataLink - Added initial data source for Zenodo
Zenodo has no meta data fields specific to diffraction image sets. There is a MX community but not all data sets are included in it.

To work around this, the following is done:
 * A list of all PDB identifiers is downloaded from wwpdb.org, unpacked and cached.
 * The data source searches via the Zenodo API for all records that have data attachments and contain the acronym "PDB".
 * From these results the title and description of each entry is matches against a regular expression that accepts PDB identifiers.
 * If there are any matches, they are checked against the list of existing PDB identifiers.
 * If this matches, then the record is added to the data source catalog

Data sets in Zenodo often have multiple files associated. Support for this was added previously in 8f85cb01

Also retry support was added in 06ab4163 which is also required for reliable downloading from Zenodo.

Zenodo does have a feature to download all record attachments in a single file, but it often times out. As it's likely created on demand, this would also not be suitable for resume support.