DataLink - remove uneeded abortJobs and simplify dl_cmdline
abortJobs was only used by dl_cmdline, to set current jobs to failed before exit.

It's cleaner to leave the tasks in an "in_progress" state, for resuming, and remove the delay on exit. The subprocesses will quit once they can process the signal that the parent process has exited.