Fix cancellation of Slurm jobs
Cancelling Slurm jobs was not working correctly due to an error with sbatch Job ID extraction and an incorrect scancel call.

The previous Job ID extraction code looped through each word in the sbatch output string and checked if it was an integer.
However the LF (\n) with the Job ID caused this to fail so the jobEntry.pid used for cancelling the job was always -1.

This is changed to:
 * strip LF and split output into an array
 * choose the last item from the array (Job ID)
 * ParseInt the Job ID and if it's an integer, set the jobEntry.pid

Remove an incorrect parameter of "kill" that was being sent to scancel that was causing the call to fail.

Note: sbatch supports a --parsable option which would simplify the code, but it isn't available with old Slurm installs so this is not used currently.